requirements.txt
gunicorn==20.1.0
djangorestframework==3.14.0
Django~=4.2.3
drf-yasg==1.21.6
